pyproject.toml
[project]
name = "the-blue-alliance"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "appengine-python-standard==2.0.0",
    "beautifulsoup4==4.14.3",
    "firebase-admin==7.2.0",
    "Flask==3.1.3",
    "Flask-Caching==2.3.1",
    "Flask-Cors==6.0.2",
    "Flask-WTF==1.2.2",
    "google-api-core==2.30.0",
    "google-api-python-client==2.192.0",
    "google-cloud-logging==3.14.0",
    "google-cloud-run==0.15.0",
    "grpcio==1.78.1",
    "gunicorn==25.1.0",
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.3",
    "numpy==2.4.3",
    "oauth2client==4.1.3",
    "openpyxl==3.1.5",
    "pyre-extensions==0.0.32",
    "pytz==2026.1.post1",
    "requests==2.32.5",
    "six==1.17.0",
    "typing-extensions==4.15.0",
    "tzlocal==5.3.1",
    "Werkzeug==3.1.6",
]

[dependency-groups]
test = [
    "freezegun==1.5.5",
    "pytest",
    "pytest-custom_exit_code",
    "pytest-split",
    "pytest-testmon",
    "requests-mock==1.12.1",
]
ci = [
    { include-group = "test" },
    "pytest-cov",
]
lint = [
    "black==26.3.0",
    "flake8==7.3.0",
    "flake8-import-order",
    "flake8-tidy-imports",
]
typecheck = [
    { include-group = "webpack" },
    "pyre-check==0.9.25",
]
webpack = [
    "watchdog==6.0.0",
]
dev = [
    { include-group = "test" },
    { include-group = "lint" },
    { include-group = "typecheck" },
    "pre-commit",
]
